Refactor: Move ResourceListModel mapping to ResourcesRepository
- Add `getResourceListModels` to `ResourcesRepository` interface.
- Move mapping logic from `ResourcesViewModel` to `ResourcesRepositoryImpl`.
- Update `ResourcesViewModel` to delegate `getLibraryListModels` to the repository.
- Update unit tests in `ResourcesViewModelTest` to correctly mock the new repository method.
